+++ Game/src/audiolib/dsl.c
@@ -221,11 +221,11 @@ unsigned DSL_GetPlaybackRate( void )
 	return _SampleRate;
 }
 
-unsigned long DisableInterrupts( void )
+uint32_t DisableInterrupts( void )
 {
 	return 0;
 }
 
-void RestoreInterrupts( unsigned long flags )
+void RestoreInterrupts( uint32_t flags )
 {
 }
